+++ src/main/java/core/basesyntax/Rectangle.java
@@ -1,10 +1,10 @@
 package core.basesyntax;
 
 public class Rectangle extends Figure {
-    private int sideA;
-    private int sideB;
+    private double sideA;
+    private double sideB;
 
-    public Rectangle(String color, int sideA, int sideB) {
+    public Rectangle(String color, double sideA, double sideB) {
         super(color);
         this.sideA = sideA;
         this.sideB = sideB;
